+++ helper-scripts/src/stackManagementTool/delete/deleteStacksTool/deleteStacks/deleteStacks.ts
