+++ helper-scripts/src/stackManagementTool/delete/deleteStacksTool/deleteStacks/prompts.ts
