+++ helper-scripts/src/stackManagementTool/delete/deleteStacksTool/deleteStacksTool.ts
