+++ helper-scripts/src/stackManagementTool/delete/deleteStacksTool/getStacks/getStacks.ts
@@ -13,7 +13,7 @@ import {
   selectingStacksErrorMessage,
   youAreAboutToDeleteMessage,
 } from "./prompts.js";
-import { protectedStacks } from "../../protectedStacks/protectedStacks.js";
+import { protectedStacks } from "../../../protectedStacks/protectedStacks.js";
 
 const getBaseStackNames = async (): Promise<string[]> => {
   const baseStackNames: string[] = [];
@@ -142,7 +142,7 @@ export const getStacks = async (): Promise<PrioritisedStacks> => {
   const candidates = await getStackCandidates(baseStackName);
   selectedStacks.push(...(await selectStacksToDelete(candidates)));
   await confirmStacks(selectedStacks);
-  await checkIfProtectedStack(selectedStacks, protectedStacks);
+  checkIfProtectedStack(selectedStacks, protectedStacks);
 
   return prioritiseStacks(selectedStacks);
 };
