+++ helper-scripts/src/stackManagementTool/delete/deleteStacksTool/getStacks/prompts.ts
