+++ helper-scripts/src/stackManagementTool/delete/selectDeletionTool/prompts.ts
