+++ helper-scripts/src/stackManagementTool/delete/selectDeletionTool/selectDeletionTool.ts
@@ -1,7 +1,7 @@
 import { echo } from "zx";
-import { deleteStacksTool } from "../../delete/deleteStacksTool.js";
-import { deleteStacksToolLegacy } from "../../deleteLegacy/deleteStacksToolLegacy.js";
+import { deleteStacksToolLegacy } from "../deleteLegacy/deleteStacksToolLegacy.js";
 import { askWhichDeletionTool } from "./prompts.js";
+import { deleteStacksTool } from "../deleteStacksTool/deleteStacksTool.js";
 
 export const whichDeletionTool = async (): Promise<void> => {
   const { choice } = await askWhichDeletionTool();
