+++ helper-scripts/src/stackManagementTool/selectTool/selectTool.ts
@@ -1,7 +1,7 @@
 import { echo } from "zx";
-import { deployDevStack } from "../deploy/deployDevStack.js";
+import { deployStackTool } from "../deploy/deployStackTool.js";
 import { askWhichTool, welcomeMessage } from "./prompts.js";
-import { whichDeletionTool } from "./selectDeletionTool/selectDeletionTool.js";
+import { whichDeletionTool } from "../delete/selectDeletionTool/selectDeletionTool.js";
 
 export const selectTool = async (): Promise<void> => {
   welcomeMessage();
@@ -12,6 +12,6 @@ export const selectTool = async (): Promise<void> => {
   if (choice === "Delete stacks") {
     await whichDeletionTool();
   } else {
-    await deployDevStack();
+    await deployStackTool();
   }
 };
